+++ internal/provider/provider.go
@@ -157,6 +157,8 @@ func (p *bridgeportProvider) Resources(_ context.Context) []func() resource.Reso
 		NewServerResource,
 		NewVarResource,
 		NewSecretResource,
+		NewConfigFragmentResource,
+		NewConfigFileResource,
 	}
 }